Split model classes into separate submodule.
To avoid circular dependencies in the resource modules.
