+++ src/commissaire/handlers/clusters.py
@@ -21,63 +21,9 @@ import falcon
 import etcd
 import json
 
-from commissaire.model import Model
 from commissaire.resource import Resource
-from commissaire.handlers.hosts import Host
 from commissaire.jobs import POOLS, clusterexec
-
-
-class Cluster(Model):
-    """
-    Representation of a Cluster.
-    """
-    _json_type = dict
-    _attributes = ('status',)
-
-    def __init__(self, **kwargs):
-        Model.__init__(self, **kwargs)
-        # Hosts is always calculated, not stored in etcd.
-        self.hosts = {'total': 0,
-                      'available': 0,
-                      'unavailable': 0}
-
-    # FIXME Generalize and move to Model?
-    def to_json_with_hosts(self):
-        data = {}
-        for key in self._attributes:
-            data[key] = getattr(self, key)
-        data['hosts'] = self.hosts
-        return json.dumps(data)
-
-
-class ClusterRestart(Model):
-    """
-    Representation of a Cluster restart operation.
-    """
-
-    _json_type = dict
-    _attributes = (
-        'status', 'restarted', 'in_process',
-        'started_at', 'finished_at')
-
-
-class ClusterUpgrade(Model):
-    """
-    Representation of a Cluster upgrade operation.
-    """
-
-    _json_type = dict
-    _attributes = (
-        'status', 'upgrade_to', 'upgraded', 'in_process',
-        'started_at', 'finished_at')
-
-
-class Clusters(Model):
-    """
-    Representation of a group of one or more Clusters.
-    """
-    _json_type = list
-    _attributes = ('clusters',)
+from commissaire.handlers.models import *
 
 
 class ClustersResource(Resource):
